Hex now includes an ImageSurface and a Context
This allows tiles to be constructed without providing a Context object.
The Context was used for determining whether a track crosses another
track, whether two tracks are connected, and whether a track intersects
a City. This can all be done using an appropriately-sized surface, which
Hex now creates and owns.