Prevent a panic when showing modal dialogs
Previously, while a modal dialog was being shown (e.g., loading or
saving a file) any attempt to redraw the map would result in a panic.
This was caused by trying to borrow the UI state in order to redraw the
map, while there was already a mutable borrow of the UI for handling the
key-press event that triggered the modal dialog.

The solution implemented here is to copy the map to another surface
every time that it is drawn, so that when the UI cannot be mutably
borrowed we can instead display the most recently-drawn content.