Consider train permutations, not combinations
This is necessary because the train ordering matters when considering
pairings of trains to paths. For example, two different trains may both
be capable of operating two different routes but the revenue earned may
differ. The previous implementation only searched over combinations and
failed to consider all possible orderings.

Note that we only need to consider permutations that result in unique
pairings of train types to routes. If a company owns two 2-trains, we
don't need to explore both permutations when evaluating the pairing of
these trains to two routes.