Apply revenue multipliers to each stop
Revenue multipliers are associated with trains that earn higher revenue
from each stop, such as the 2+2 and 5+5E trains in 1861 and 1867.
Previously, the revenue multiplier was applied to the net revenue earned
from the entire route, and so the revenue recorded at each visit did not
reflect the true amount of revenue earned.