Improve the route-finding UI for rusty_train
The user is now prompted to select a company from those that have placed
tokens on the map, rather than selecting a tile and then selecting a
token space.

This adds the method Map::unique_placed_tokens(), which returns the set
of tokens that have been placed on the map.

I have chosen to return these tokens in a BTreeSet, so that the order is
consistent. This, in turn, means that several data types now implement
PartialOrd and Ord.