Print invalid token names to stderr
This can occur when placing tiles with Example::place_tiles(), and when
loading or updating maps with Descr::build_map() or Descr::update_map().

Note that encountering invalid token names in these situations still
causes a panic, but now it is possible to identify the token name that
triggered the panic.