Define a default hex for constructing tiles
This ensures that track connectivity is as expected, regardless of the
hexagon size and theme with which they are drawn. It also allows the hex
argument to be removed from various functions, such as n18io::read_tile
and n18io::read_tiles, and means that n18game::_1867::Game can implement
Default.