Make types implement Ord, but not Hash
Values of these types can be stored in B-Tree maps and B-Tree sets, but
not in hash-based collections such as HashMap and HashSet.

This ensures that iteration over these types has a well-defined ordering
and that the results are deterministic. While it is possible to override
the default hashing algorithm with a deterministic alternative, it is
simpler to enforce the use of collections that rely on Ord. Doing so has
also resulted in a small but consistent increase in performance, as
noted in the Performance chapter of the Developer guide.

Note that some types cannot implement Ord because they contain floating
point values. This applies to n18hex::Coord, n18tile::Track, and the
private type n18tile::track::TrackPath.