Decouple the user interface from the current state
This restructures the entire n18ui crate, dividing the user interface
into several components and separating the user interface states from
the event-handling logic.

It also fixes how tile upgrades and replacements are drawn, so that the
candidate tile and the placed tokens are correctly oriented regardless
of the underlying tile's rotation.

This fix changes how n18brush::draw_tile_and_tokens_at() interprets its
`radians` argument; it is now the rotation that is applied in addition
to the currently-placed tile's rotation (if any), rather than being the
net rotation.