Draw off-board track as triangular segments
Use Tile::with_offboard_faces() to define the faces for which such track
segments should be drawn.

This changes how off-board track segments are drawn, but does not modify
the underlying tile contents (e.g., track segments, dits, and cities).
It updates the n18brush crate so that paths and routes are highlighted
correctly when they include off-board track and cities.