Fix how 1867_bc positions net revenue labels
The 1867_bc example used Labeller::draw_at_current_point() and relied on
there being no current point so that the label would be drawn at the
origin (0, 0). This in turn meant that n18brush::highlight_visits() had
to ensure there was no current point when it returned. Now the 1867_bc
example explicitly positions the labels at the origin, and there is no
need for n18brush::highlight_visits() to create unnecessary paths.