Simplify the n18map::Map internal state
This replaces two BTreeMap fields and one Vec field with one BTreeMap,
and introduces some minor changes to Map's public methods. It prevents
issues with duplicate hex addresses, and ensures that hex addresses and
placed tiles are iterated over in sorted order, regardless of the order
in which they were defined when passed to Map::new().