Avoid recursive main loops when showing dialogs
This is a relatively invasive change in n18ui. Both the main UI struct
and various State objects use dialogs to prompt the user, and so we need
to distinguish between "pings" intended for the current State, and those
intended for the main UI struct.

The FindRoutes state is now divided into four separate states, which
directly reflect the sequence of dialog prompts and route-searching.