Implement 1830: Railways and Robber Barons
Having learned from the implementation of 1861 and 1867, the _1830
module is better organised and is divided into several sub-modules.

This adds 15 new tiles to n18catalogue: 1, 2, 53-56, 59-62, 64-69.

The save_tiles and draw_tiles examples now include all 1830 tiles, and
the draw_tiles example now supports both hex orientations, so it can
draw the 1830 tiles with their correct orientation.

This also updates the default hex theme: the location label width has
been increased from 80.0 to 85.0, so that "Washington" can be displayed
in a single line. With the old width, the final "n" was wrapped.
This results in minor changes in several tiles for 1861 and 1867, and
upgrading from Pango 1.46.2 to 1.48.10 has also introduced minor kerning
changes in several tiles.