Avoid a panic in n18route::search::paths_from()
When the starting point for the search has no connections, this now
returns an empty vector rather than panicking. Note that this is an
unlikely set of circumstances, since it would mean that a token has been
placed in a token space that is not connected to any track segments.